# JaneStreet profile: treat comments as doc-comments