--- a/lib/Conf_t.ml
+++ b/lib/Conf_t.ml
@@ -118,6 +118,7 @@ type fmt_opts =
   ; type_decl: [`Compact | `Sparse] elt
   ; type_decl_indent: int elt
   ; wrap_comments: bool elt
+  ; wrap_docstrings: bool elt
   ; wrap_fun_args: bool elt }
 
 type opr_opts =