--- a/lib/Conf_t.mli
+++ b/lib/Conf_t.mli
@@ -116,6 +116,7 @@ type fmt_opts =
   ; type_decl: [`Compact | `Sparse] elt
   ; type_decl_indent: int elt
   ; wrap_comments: bool elt  (** Wrap comments at margin. *)
+  ; wrap_docstrings: bool elt
   ; wrap_fun_args: bool elt }
 
 (** Options changing the tool's behavior *)