--- a/test/passing/tests/doc_comments.mli.ref
+++ b/test/passing/tests/doc_comments.mli.ref
@@ -253,7 +253,9 @@ end
 
 (** @canonical Module.Foo.Bar *)
 
-(** {v a v} *)
+(** {v
+a
+    v} *)
 
 (** {[
       b