--- a/test/passing/tests/js_sig.mli.err
+++ b/test/passing/tests/js_sig.mli.err
@@ -1,0 +1,1 @@
+Warning: tests/js_sig.mli:11 exceeds the margin