--- a/test/passing/tests/js_sig.mli.ref
+++ b/test/passing/tests/js_sig.mli.ref
@@ -9,12 +9,10 @@ exception Second_exception
 [@@@ocamlformat "parse-docstrings=true"]
 [@@@ocamlformat "wrap-comments=true"]
 
-(** {e foooooooo oooooo oooo oooo ooooo oooo ooooo}
-    {i fooooo ooooo ooo oooooo oo oooooo oooo}
-    {b fooooooo oooooo oooooo oooooo oooooo ooooooo} *)
+(** {e foooooooo oooooo oooo oooo ooooo oooo ooooo} {i fooooo ooooo ooo oooooo oo oooooo oooo} {b fooooooo oooooo oooooo oooooo oooooo ooooooo} *)
 
-(** {e foooooooo oooooooooo ooooooooo ooooooooo} {{!some ref} fooooooooooooo oooooooo
-    oooooooooo} {b fooooooooooooo oooooooooooo oooooo ooooooo} *)
+(** {e foooooooo oooooooooo ooooooooo ooooooooo} {{!some ref} fooooooooooooo
+    oooooooo oooooooooo} {b fooooooooooooo oooooooooooo oooooo ooooooo} *)
 
 class c : 'a -> object
   val x : 'b