--- a/test/passing/tests/js_source.ml
+++ b/test/passing/tests/js_source.ml
@@ -7778,3 +7778,6 @@ end = struct
   external unsafe_memset : t -> pos:int -> len:int -> char -> unit = "bigstring_memset_stub"
   [@@noalloc]
 end
+
+(*
+   *)