--- a/test/passing/tests/js_source.ml
+++ b/test/passing/tests/js_source.ml
@@ -7781,3 +7781,9 @@ end
 
 (*
    *)
+
+(**
+   xxx
+*)
+include S1
+(** @inline *)