--- a/test/passing/tests/js_source.ml.err
+++ b/test/passing/tests/js_source.ml.err
@@ -1,7 +1,7 @@
 Warning: tests/js_source.ml:155 exceeds the margin
 Warning: tests/js_source.ml:3554 exceeds the margin
-Warning: tests/js_source.ml:9515 exceeds the margin
-Warning: tests/js_source.ml:9618 exceeds the margin
-Warning: tests/js_source.ml:9637 exceeds the margin
-Warning: tests/js_source.ml:9677 exceeds the margin
-Warning: tests/js_source.ml:9759 exceeds the margin
+Warning: tests/js_source.ml:9514 exceeds the margin
+Warning: tests/js_source.ml:9617 exceeds the margin
+Warning: tests/js_source.ml:9636 exceeds the margin
+Warning: tests/js_source.ml:9676 exceeds the margin
+Warning: tests/js_source.ml:9760 exceeds the margin