--- a/test/passing/tests/js_source.ml.ocp
+++ b/test/passing/tests/js_source.ml.ocp
@@ -2863,8 +2863,8 @@ let rec pre_subst f = function
 ;;
 
 let comp_subst f g (x : 'a fin) = pre_subst f (g x)
-(*  val comp_subst :
-    ('b fin -> 'c term) -> ('a fin -> 'b term) -> 'a fin -> 'c term *)
+(* val comp_subst :
+   ('b fin -> 'c term) -> ('a fin -> 'b term) -> 'a fin -> 'c term *)
 
 (* 4 The Occur-Check, through thick and thin *)
 
@@ -2959,7 +2959,7 @@ let rec sub' : type m. m ealist -> m fin -> m term = function
 ;;
 
 let subst' d = pre_subst (sub' d)
-(*  val subst' : 'a ealist -> 'a term -> 'a term *)
+(* val subst' : 'a ealist -> 'a term -> 'a term *)
 
 (* 6 First-Order Unification *)
 
@@ -7471,7 +7471,6 @@ end
    module
    rec NewA : ASig = MakeA (struct end)
    and NewB : BSig with type b = NewA.a = MakeB (struct end);;
-
 *)
 
 (* Expressions and bindings *)
@@ -7872,7 +7871,7 @@ end =
 
 let _ = test 103 (Lazy.force M3.x) 3
 
-(** Pure type-checking tests: see recmod/*.ml  *)
+(** Pure type-checking tests: see recmod/*.ml *)
 type t =
   | A of
       { x : int
@@ -9690,14 +9689,16 @@ type t =
 type t =
   { field : ty
   (* Here is some verbatim formatted text:
+
      {v
        starting at column 7
-     v}*)
+         v}*)
   }
 
 module Intro_sort = struct
   let foo_fooo_foooo fooo ~foooo m1 m2 m3 m4 m5 =
     (* Fooooooooooooooooooooooooooo:
+
        {v
           1--o-----o-----o--------------1
              |     |     |
@@ -9708,7 +9709,7 @@ module Intro_sort = struct
           4-----o--------o--o--|-----o--4
                 |              |     |
           5-----o--------------o-----o--5
-        v} *)
+       v} *)
     foooooooooo fooooo fooo;
     foooooooooo fooooo fooo;
     foooooooooo fooooo fooo
@@ -10008,3 +10009,6 @@ end = struct
     = "bigstring_memset_stub"
   [@@noalloc]
 end
+
+(*
+*)