--- a/test/passing/tests/js_source.ml.ref
+++ b/test/passing/tests/js_source.ml.ref
@@ -10012,3 +10012,7 @@ end
 
 (*
    *)
+
+(** xxx *)
+include S1
+(** @inline *)