@gpetiot 2023-03-07
@ceastlund: does this new behavior look good to you?