@ceastlund 2023-03-07
I'm not quite sure I understand the change here. I believe ocp-indent may change the indentation of the opening `{v`, since the whitespace preceding it is not verbatim. But anything between `{v` and `v}` should not be changed, right? That should be the purpose of a verbatim block.