@ceastlund 2023-03-07
Oh, I see, you're trying to move stuff relative to the comment itself. I'd say don't bother with that. Treat `{v ... v}` like `{| ... |}` strings, if you can: you can move the starting points, but the contents don't shift, even if that makes them unaligned with surrounding code.