@Julow 2023-03-09
You are right, according to Odoc's parser (that is not used here but I think make sense to mention), the content of a verbatim shouldn't change, even its indentation.